Downloads/agentic-auto-job-applier-nodocker/requirements.txt
streamlit
selenium
beautifulsoup4
requests
sentence-transformers
python-dotenv
pdfminer.six
docx2txt